requirements.txt
python-dotenv
llama-index
llama-index-tools-linkup-research
streamlit 
qdrant-client

llama-index-llms-ollama

llama-index-llms-azure-openai
llama-index-vector-stores-qdrant
llama-index-embeddings-fastembed
openai